JETT-3840
Updated the Utility method -- mapStoreConfigs to map the store to the
Active Store Config rather than any store config linked to the store.
This method was originally written before we had the Active Store Config
lookup on the Store and assumed that there would only be one Store
Config linked to a store. However, if that is not the case, the method
has the potential to pull the wrong store config. Updated the code in
multiple places to use the updated method.