JETT-4229
Updated the vendorpurchaseordercontrollers so that When a Service
Product of "Complete Unit" is received, the asset created needs to have
the original order item product name (master product) but with the new
variant number, so it will allow service products to be tied to it when
creating a future service request. 
Move the replaced Installed product to "Replaced" status.